Removed code generation and nodes and now rendered instead of compiled

+++ src/main/java/com/mitchellbosecke/pebble/extension/core/EvenTest.java
@@ -18,7 +18,7 @@ public class EvenTest implements Test {
 			throw new IllegalArgumentException("Can not pass null value to \"even\" test.");
 		}
 
-		Integer obj = (Integer) input;
+		Long obj = (Long) input;
 		return (obj % 2 == 0);
 	}
 }
